Code edited

+++ src/BmiService.java
@@ -1,5 +1,5 @@
 public class BmiService {
-    public double calculate(int weightkg, int heightcm) {
+    public double calculate(int weightkg, float heightcm) {
 
         double bmi = (double) weightkg /  ((heightcm * heightcm) / 100);
